Moving determineTypeFromClass from Type interface to ParserHelper class.
