+++ kie-dmn-feel/src/main/java/org/kie/dmn/feel/parser/feel11/ParserHelper.java
@@ -22,18 +22,34 @@ import org.antlr.v4.runtime.misc.Interval;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.TerminalNode;
 import org.kie.dmn.feel.lang.CustomType;
+import org.kie.dmn.feel.lang.FEELAccessor;
 import org.kie.dmn.feel.lang.CustomType.Field;
+import org.kie.dmn.feel.lang.impl.JavaBackedType;
 import org.kie.dmn.feel.lang.Scope;
 import org.kie.dmn.feel.lang.Symbol;
 import org.kie.dmn.feel.lang.Type;
+import org.kie.dmn.feel.lang.types.BuiltInType;
 import org.kie.dmn.feel.lang.types.ScopeImpl;
 import org.kie.dmn.feel.lang.types.SymbolTable;
 import org.kie.dmn.feel.lang.types.VariableSymbol;
+import org.kie.dmn.feel.runtime.FEELFunction;
+import org.kie.dmn.feel.runtime.Range;
+import org.kie.dmn.feel.runtime.UnaryTest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.time.Duration;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.OffsetDateTime;
+import java.time.OffsetTime;
+import java.time.Period;
+import java.time.ZonedDateTime;
 import java.util.List;
+import java.util.Map;
 import java.util.Stack;
+import java.util.stream.Stream;
 
 public class ParserHelper {
     public static final Logger LOG = LoggerFactory.getLogger(ParserHelper.class);
@@ -147,8 +163,37 @@ public class ParserHelper {
         return tokens;
     }
 
-    public void dump() {
-        LOG.trace("dump"+currentName);
+    public static Type determineTypeFromClass( Class<?> clazz ) {
+        if( clazz == null ) {
+            return BuiltInType.UNKNOWN;
+        } else if( Number.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.NUMBER;
+        } else if( String.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.STRING;
+        } else if( LocalDate.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.DATE;
+        } else if( LocalTime.class.isAssignableFrom(clazz) || OffsetTime.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.TIME;
+        } else if( ZonedDateTime.class.isAssignableFrom(clazz) || OffsetDateTime.class.isAssignableFrom(clazz) || LocalDateTime.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.DATE_TIME;
+        } else if( Duration.class.isAssignableFrom(clazz) || Period.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.DURATION;
+        } else if( Boolean.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.BOOLEAN;
+        } else if( UnaryTest.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.UNARY_TEST;
+        } else if( Range.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.RANGE;
+        } else if( FEELFunction.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.FUNCTION;
+        } else if( List.class.isAssignableFrom(clazz) ) {
+            return BuiltInType.LIST;
+        } else if( Map.class.isAssignableFrom(clazz) ) {     // TODO not so sure about this one..
+            return BuiltInType.CONTEXT;
+        } else if (Stream.of(clazz.getMethods()).anyMatch(m->m.getAnnotation(FEELAccessor.class)!=null)) {
+            return new JavaBackedType(clazz);
+        }
+        return BuiltInType.UNKNOWN;
     }
 
 }
